+++ src/SysWatt.App/Views/SettingsWindow.xaml.cs
@@ -1,4 +1,5 @@
 using System.Windows;
+using System.Windows.Input;
 using SysWatt.App.ViewModels;
 
 namespace SysWatt.App.Views;
@@ -14,4 +15,11 @@ public partial class SettingsWindow : Window
     }
 
     private void ExportDiagnostics_Click(object sender, RoutedEventArgs e) => ExportDiagnosticsRequested?.Invoke(this, EventArgs.Empty);
+
+    private void Header_MouseLeftButtonDown(object sender, MouseButtonEventArgs e)
+    {
+        if (e.LeftButton == MouseButtonState.Pressed) DragMove();
+    }
+
+    private void Close_Click(object sender, RoutedEventArgs e) => Close();
 }